# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AppCategory = typeof AppCategory[keyof typeof AppCategory];


export const AppCategory = {
  Productivity: 'Productivity',
  Entertainment: 'Entertainment',
  Utilities: 'Utilities',
  Creative: 'Creative',
} as const;

export interface App {
  id: string;
  name: string;
  publisher: string;
  category: AppCategory;
  description: string;
  detail: string;
  price: number;
  size: string;
  version: string;
  initials: string;
  iconBg: string;
  /** @nullable */
  iconFg?: string | null;
  /** @nullable */
  imageDataUrl?: string | null;
}

export type AppInputCategory = typeof AppInputCategory[keyof typeof AppInputCategory];


export const AppInputCategory = {
  Productivity: 'Productivity',
  Entertainment: 'Entertainment',
  Utilities: 'Utilities',
  Creative: 'Creative',
} as const;

export interface AppInput {
  id: string;
  name: string;
  publisher: string;
  category: AppInputCategory;
  description: string;
  detail: string;
  price: number;
  size: string;
  version: string;
  initials: string;
  iconBg: string;
  /** @nullable */
  iconFg?: string | null;
  /** @nullable */
  imageDataUrl?: string | null;
}