# lib/api-zod/src/generated/types/appInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppInputCategory } from './appInputCategory';

export interface AppInput {
  id: string;
  name: string;
  publisher: string;
  category: AppInputCategory;
  description: string;
  detail: string;
  price: number;
  size: string;
  version: string;
  initials: string;
  iconBg: string;
  /** @nullable */
  iconFg?: string | null;
  /** @nullable */
  imageDataUrl?: string | null;
}